+++ scipy/optimize/optimize.py
@@ -610,7 +610,7 @@ def approx_fprime(xk, f, epsilon, *args):
 
     >>> x = np.ones(2)
     >>> c0, c1 = (1, 200)
-    >>> eps = np.sqrt(np.finfo(np.float).eps)
+    >>> eps = np.sqrt(np.finfo(float).eps)
     >>> optimize.approx_fprime(x, func, [eps, np.sqrt(200) * eps], c0, c1)
     array([   2.        ,  400.00004198])
